Remove the unlikely case in loadi function (saved 4 instructions)
One day, I revisit this file and noticed the two debug lines that is supposed to print out the result when `loadi` loads an instructions that is split on 2 different sectors.

These 2 lines have existed in `loadi` from the very beginning of the function, but somehow didn't print anything in all of my tests. After some research, I believe it's due to the way RISC-V instructions are aligned on 32-bit boundaries, so it will never be split on 2 different 512-bytes sector.

So the whole `if` statement is removed. This saves 4 instructions in runtime and even more when considering the programming space. I wonder how much will the performance improve?